Move camera's logic into its own file
