+++ src/client.rs
@@ -1,12 +1,13 @@
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
-use crate::settings::NetworkSettings;
-use crate::shared::shared_config;
-use bevy::{prelude::*, render::view::RenderLayers};
-
+use bevy::prelude::*;
 use client::*;
 use lightyear::prelude::*;
 
+use crate::settings::NetworkSettings;
+use crate::shared::shared_config;
+
+mod camera;
 mod lobby;
 mod player;
 mod ui;
@@ -21,15 +22,19 @@ impl Plugin for ClientPlugin {
         let settings = app.world().get_resource::<NetworkSettings>().unwrap();
         app.add_plugins(ClientPlugins::new(client_config(settings)));
 
-        app.add_plugins((lobby::LobbyPlugin, ui::UiPlugin, player::PlayerPlugin))
-            .init_state::<Connection>()
-            .enable_state_scoped_entities::<Connection>()
-            .add_systems(Startup, spawn_game_camera)
-            .add_systems(OnEnter(Connection::Connect), connect_server)
-            .add_systems(
-                PreUpdate,
-                (handle_connection, handle_disconnection).after(MainSet::Receive),
-            );
+        app.add_plugins((
+            lobby::LobbyPlugin,
+            ui::UiPlugin,
+            player::PlayerPlugin,
+            camera::CameraPlugin,
+        ))
+        .init_state::<Connection>()
+        .enable_state_scoped_entities::<Connection>()
+        .add_systems(OnEnter(Connection::Connect), connect_server)
+        .add_systems(
+            PreUpdate,
+            (handle_connection, handle_disconnection).after(MainSet::Receive),
+        );
 
         // Enable dev tools for dev builds.
         #[cfg(feature = "dev")]
@@ -65,21 +70,6 @@ fn handle_disconnection(
 
         connection.set(Connection::Disconnected);
     }
-}
-
-/// Spawn camera for game rendering (render layer 0).
-fn spawn_game_camera(mut commands: Commands) {
-    commands.spawn((
-        Name::new("Game Camera"),
-        Camera2dBundle {
-            camera: Camera {
-                clear_color: Color::NONE.into(),
-                ..default()
-            },
-            ..default()
-        },
-        RenderLayers::layer(0),
-    ));
 }
 
 /// Create the lightyear [`ClientConfig`].
